refactor(adapter-openclaw): extract channels + messaging into channel-helpers.ts
Ninth cut of the runtime.ts god-file split. Moves the 10 channel/messaging
helpers into a new channel-helpers.ts: channel ref/arg building
(splitChannelRef, openClawMessageSendArgs), delivery-output parsing
(deliveryRefFromOpenClawOutput, parseOpenClawDeliveryOutput), the channel
registry read + interactive-approval detection (readChannelInfos,
hasAnyInteractiveApprovalChannel, channelHasInteractiveApproval,
channelEntrySupportsInteractiveApproval, humanizeChannelId), and
approvalNoticeForMessage (the channel-coupled approval-notice text deferred
from the approval-helpers cut) + its NATIVE_APPROVAL_PROVIDERS /
RENDER_ONLY / REJECT_REASON constants + OPENCLAW_PLUGIN_APPROVAL_TIMEOUT_MS.

approvalNoticeForMessage imports requiresRejectReason from approval-helpers
(sibling). gatewayWebSocketUrl stays on the adapter — it is coupled to the
OpenClawSettings type, which the class owns. The class's channels methods
import back the 7 helpers + the timeout const they call; the rest stay
module-internal. runtime.ts shed its now-unused firstStringAtPaths/isRecord
imports.

Pure relocation — bodies byte-for-byte unchanged. runtime.ts 1,744 -> 1,618.
Verified: typecheck, eslint (changed files), adapter suite 115-0 incl.
runtime-channels.test.ts (--isolate), full suite 5124-0, binary build (stamps
reverted), boot smoke (adapter loads, schedule activates, list route 200).
